requirements.txt
# Career Ops Suite — dependencies
# Pinned for reproducibility; compatible with Streamlit Cloud free tier.
# No system binaries required.

# Web framework
streamlit>=1.35.0,<2.0.0

# Anthropic Claude API
anthropic>=0.28.0,<1.0.0

# Markdown processing
markdown>=3.6,<4.0

# PDF generation (pure Python, no system dependencies)
xhtml2pdf>=0.2.11,<0.3.0

# Environment variable loading
python-dotenv>=1.0.0,<2.0.0
